refactor: import statements and add relationship module
- Updated import statements in multiple files to catch ImportError specifically.
- Added a new relationship module extracted from the legacy codebase.
- Ensured consistent formatting and organization of imports across various reader and writer modules.
