+++ src/pyArchimate/helpers/__init__.py
@@ -1,0 +1,18 @@
+"""Helper modules for diagram, property, and logging utilities."""
+
+from .diagram import get_or_create_connection, get_or_create_node
+from .logging import log, log_set_level, log_to_file, log_to_stderr
+from .properties import check_invalid_conn, check_invalid_nodes, embed_props, expand_props
+
+__all__ = [
+    "get_or_create_connection",
+    "get_or_create_node",
+    "check_invalid_conn",
+    "check_invalid_nodes",
+    "embed_props",
+    "expand_props",
+    "log",
+    "log_set_level",
+    "log_to_file",
+    "log_to_stderr",
+]
