+++ src/pyArchimate/helpers/logging.py
@@ -1,0 +1,8 @@
+"""Logging helpers re-exported from the package logger module."""
+
+try:
+    from ..logger import log, log_set_level, log_to_file, log_to_stderr
+except ImportError:
+    from logger import log, log_set_level, log_to_file, log_to_stderr
+
+__all__ = ["log", "log_set_level", "log_to_file", "log_to_stderr"]
