+++ src/pyArchimate/model.py
@@ -1,0 +1,8 @@
+"""Model module extracted from the legacy monolith."""
+
+try:
+    from ._legacy import Model
+except ImportError:
+    from _legacy import Model
+
+__all__ = ["Model"]
