+++ src/pyArchimate/readers/archimateReader.py
@@ -1,8 +1,8 @@
 import sys
-import lxml.etree as et
+
 try:
     from .. import *
-except:
+except ImportError:
     sys.path.insert(0, "..")
     from pyArchimate import *
