+++ src/pyArchimate/readers/arisAMLreader.py
@@ -9,10 +9,10 @@
 import ctypes
 import platform
 import sys
-import lxml.etree as et
+
 try:
     from .. import *
-except:
+except ImportError:
     sys.path.insert(0, "..")
     from pyArchimate import *
