+++ src/pyArchimate/writers/archimateWriter.py
@@ -2,7 +2,7 @@ import sys
 
 try:
     from .. import *
-except:
+except ImportError:
     sys.path.insert(0, "..")
     from pyArchimate import *
